twitter-bot-python/requirements.txt
# Function dependencies, for example:
# package>=version

requests==2.18.3
requests-oauthlib==1.3.0
python-dotenv==0.20.0
schedule==1.1.0
